Speeding up gameplay
With every call to Update(), the speed of the AppleTree and the rate of apple drops increases. I think I did this pretty inelegantly, but I'm not sure what to use to make it better, yet. I wanted to try it on my own before I looked up the best way to do it. Also not sure if it's better to have this in Update(), FixedUpdate(), or LateUpdate().