Why are you doing that?
I mostly figured out how to get the apple to explode. I added a plane to act as a floor so the apple doesn't keep falling and the ApplePicker script now destroys all the falling apples, except the one that's exploding so the particle system will run to its completion. Now though, the scene resets when an apple is missed, and I have no idea why...