Oh my god, I did it. Re: Particle Systems
Stupid exploding apple works perfectly. Took me like there days to figure it out, but I did it! So the way it works is that if an apple is missed (hits the bottomY value in Apple.cs) the applePicker script will delete all the apples, except the one exploding and will remove a basket. Once the partical system is done playing, that missed apple is then destroyed.